Add files via upload
In the Parental Control web browser, we have used multiple methods to protect our youth from offensive contents. The method which we are going to implement on our project is listed below:
a.	The first method is mode selection where users are divided into three categories e.g.: Parent, Child and Monitor.
b.	The second method we will going to implement is firewall.
c.	The third method which we are also going to implement is history panel to monitor child’s activities.
d.	The fourth method for data storing we will be going to implement text filling concept through which we secure our data from losing